Renamed eservice-template-updater consumer
