+++ packages/eservice-template-instances-updater/.env
@@ -8,7 +8,7 @@ ESERVICE_TEMPLATE_TOPIC="event-store.eservice-template.events"
 AWS_REGION="eu-south-1"
 
 INTERNAL_JWT_KID=ffcc9b5b-4612-49b1-9374-9d203a3834f2
-INTERNAL_JWT_SUBJECT="dev-refactor.interop-eservice-template-updater"
+INTERNAL_JWT_SUBJECT="dev-refactor.interop-eservice-template-instances-updater"
 INTERNAL_JWT_ISSUER="dev-refactor.interop.pagopa.it"
 INTERNAL_JWT_AUDIENCE="refactor.dev.interop.pagopa.it/internal"
 INTERNAL_JWT_SECONDS_DURATION=60
