+++ packages/eservice-template-instances-updater/Dockerfile
@@ -8,7 +8,7 @@ COPY pnpm-lock.yaml /app/
 COPY pnpm-workspace.yaml /app/
 COPY .npmrc /app/
 
-COPY ./packages/eservice-template-updater/package.json /app/packages/eservice-template-updater/package.json
+COPY ./packages/eservice-template-instances-updater/package.json /app/packages/eservice-template-instances-updater/package.json
 COPY ./packages/commons/package.json /app/packages/commons/package.json
 COPY ./packages/models/package.json /app/packages/models/package.json
 COPY ./packages/kafka-iam-auth/package.json /app/packages/kafka-iam-auth/package.json
@@ -18,7 +18,7 @@ RUN --mount=type=cache,id=pnpm,target=/pnpm/store pnpm install --frozen-lockfile
 
 COPY tsconfig.json /app/
 COPY turbo.json /app/
-COPY ./packages/eservice-template-updater /app/packages/eservice-template-updater
+COPY ./packages/eservice-template-instances-updater /app/packages/eservice-template-instances-updater
 COPY ./packages/commons /app/packages/commons
 COPY ./packages/models /app/packages/models
 COPY ./packages/kafka-iam-auth /app/packages/kafka-iam-auth
@@ -29,20 +29,20 @@ RUN pnpm build && \
     rm -rf /app/node_modules/.cache && \
     mkdir /out && \
     cp -a --parents -t /out \
-      node_modules packages/eservice-template-updater/node_modules \
-      package*.json packages/eservice-template-updater/package*.json \
+      node_modules packages/eservice-template-instances-updater/node_modules \
+      package*.json packages/eservice-template-instances-updater/package*.json \
       packages/commons \
       packages/models \
       packages/kafka-iam-auth \
       packages/api-clients \
-      packages/eservice-template-updater/dist && \
+      packages/eservice-template-instances-updater/dist && \
     find /out -exec touch -h --date=@0 {} \;
 
 FROM node:20.14.0-slim@sha256:5e8ac65a0231d76a388683d07ca36a9769ab019a85d85169fe28e206f7a3208e as final 
 
 COPY --from=build /out /app
 
-WORKDIR /app/packages/eservice-template-updater
+WORKDIR /app/packages/eservice-template-instances-updater
 EXPOSE 3000
 
 CMD ["node", "."]
